requirements.txt

click==8.1.2
colorama==0.4.4
Flask==2.1.1
gunicorn==20.0.4
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.26.4
pandas>=1.5.0
Pillow
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.8.0
torchvision==0.23.0
typing_extensions>=4.10.0
Werkzeug==2.1.1
zipp==3.8.0
pymongo
bcrypt
